+++ packages/config/src/assets.ts
@@ -35,11 +35,6 @@ export const astr = new Asset({
   originSymbol: 'ASTR',
 });
 
-export const atom = new Asset({
-  key: 'atom',
-  originSymbol: 'ATOM',
-});
-
 export const auq = new Asset({
   key: 'auq',
   originSymbol: 'AUQ',
@@ -115,6 +110,11 @@ export const eth = new Asset({
   originSymbol: 'ETH',
 });
 
+export const eurc = new Asset({
+  key: 'eurc',
+  originSymbol: 'EURC',
+});
+
 export const fil = new Asset({
   key: 'fil',
   originSymbol: 'FIL',
@@ -163,11 +163,6 @@ export const kbtc = new Asset({
 export const kint = new Asset({
   key: 'kint',
   originSymbol: 'KINT',
-});
-
-export const kma = new Asset({
-  key: 'kma',
-  originSymbol: 'KMA',
 });
 
 export const ksm = new Asset({
@@ -428,7 +423,6 @@ export const assetsList: Asset[] = [
   apillon,
   aseed,
   astr,
-  atom,
   auq,
   axlusdc,
   betaDEV,
@@ -444,6 +438,7 @@ export const assetsList: Asset[] = [
   eq,
   eqd,
   eth,
+  eurc,
   fil,
   ftm,
   ftmwh,
@@ -454,7 +449,6 @@ export const assetsList: Asset[] = [
   kar,
   kbtc,
   kint,
-  kma,
   ksm,
   laos,
   ldot,