+++ packages/config/src/xcm-configs/moonriver.ts
@@ -12,11 +12,9 @@ import {
   kar,
   kbtc,
   kint,
-  kma,
   ksm,
   mgx,
   movr,
-  pica,
   rmrk,
   sdn,
   teer,
@@ -29,7 +27,6 @@ import {
 } from '../assets';
 import {
   bifrostKusama,
-  calamari,
   crustShadow,
   darwiniaCrab,
   integritee,
@@ -39,7 +36,6 @@ import {
   kusamaAssetHub,
   mangataKusama,
   moonriver,
-  picasso,
   robonomics,
   shiden,
   turing,
@@ -68,26 +64,6 @@ export const moonriverRoutes = new ChainRoutes({
           asset: movr,
         },
         min: AssetMinBuilder().assetRegistry().currencyMetadatas(),
-      },
-      contract: ContractBuilder().Xtokens().transfer(),
-    },
-    {
-      source: {
-        asset: movr,
-        balance: BalanceBuilder().substrate().system().account(),
-        destinationFee: {
-          balance: BalanceBuilder().substrate().system().account(),
-        },
-      },
-      destination: {
-        asset: movr,
-        chain: calamari,
-        balance: BalanceBuilder().substrate().assets().account(),
-        fee: {
-          amount: 0.001,
-          asset: movr,
-        },
-        min: AssetMinBuilder().assets().asset(),
       },
       contract: ContractBuilder().Xtokens().transfer(),
     },
@@ -184,25 +160,6 @@ export const moonriverRoutes = new ChainRoutes({
         balance: BalanceBuilder().substrate().tokens().accounts(),
         fee: {
           amount: 0.004,
-          asset: movr,
-        },
-      },
-      contract: ContractBuilder().Xtokens().transfer(),
-    },
-    {
-      source: {
-        asset: movr,
-        balance: BalanceBuilder().substrate().system().account(),
-        destinationFee: {
-          balance: BalanceBuilder().substrate().system().account(),
-        },
-      },
-      destination: {
-        asset: movr,
-        chain: picasso,
-        balance: BalanceBuilder().substrate().tokens().accounts(),
-        fee: {
-          amount: 0.001,
           asset: movr,
         },
       },
@@ -374,29 +331,6 @@ export const moonriverRoutes = new ChainRoutes({
     },
     {
       source: {
-        asset: kma,
-        balance: BalanceBuilder().substrate().assets().account(),
-        fee: {
-          asset: movr,
-          balance: BalanceBuilder().substrate().system().account(),
-        },
-        destinationFee: {
-          balance: BalanceBuilder().substrate().assets().account(),
-        },
-      },
-      destination: {
-        asset: kma,
-        chain: calamari,
-        balance: BalanceBuilder().substrate().system().account(),
-        fee: {
-          amount: 0.000004,
-          asset: kma,
-        },
-      },
-      contract: ContractBuilder().Xtokens().transfer(),
-    },
-    {
-      source: {
         asset: ksm,
         balance: BalanceBuilder().substrate().assets().account(),
         fee: {
@@ -440,29 +374,6 @@ export const moonriverRoutes = new ChainRoutes({
         fee: {
           amount: 5.5,
           asset: mgx,
-        },
-      },
-      contract: ContractBuilder().Xtokens().transfer(),
-    },
-    {
-      source: {
-        asset: pica,
-        balance: BalanceBuilder().substrate().assets().account(),
-        fee: {
-          asset: movr,
-          balance: BalanceBuilder().substrate().system().account(),
-        },
-        destinationFee: {
-          balance: BalanceBuilder().substrate().assets().account(),
-        },
-      },
-      destination: {
-        asset: pica,
-        chain: picasso,
-        balance: BalanceBuilder().substrate().system().account(),
-        fee: {
-          amount: 0.001,
-          asset: pica,
         },
       },
       contract: ContractBuilder().Xtokens().transfer(),