+++ packages/config/src/xcm-configs/pendulum.ts
@@ -1,9 +1,10 @@
 import {
+  AssetMinBuilder,
   BalanceBuilder,
   ExtrinsicBuilder,
   FeeBuilder,
 } from '@moonbeam-network/xcm-builder';
-import { axlusdc, glmr, pen } from '../assets';
+import { axlusdc, eurc, glmr, pen } from '../assets';
 import { moonbeam, pendulum } from '../chains';
 import { ChainRoutes } from '../types/ChainRoutes';
 
@@ -80,6 +81,31 @@ export const pendulumRoutes = new ChainRoutes({
         },
       },
       extrinsic: ExtrinsicBuilder().xTokens().transferMultiCurrencies(),
+    },
+    {
+      source: {
+        asset: eurc,
+        balance: BalanceBuilder().substrate().tokens().accounts(),
+        fee: {
+          asset: eurc,
+          balance: BalanceBuilder().substrate().system().account(),
+        },
+        destinationFee: {
+          balance: BalanceBuilder().substrate().tokens().accounts(),
+        },
+        min: AssetMinBuilder().assetRegistry().metadata(),
+      },
+      destination: {
+        asset: eurc,
+        chain: moonbeam,
+        balance: BalanceBuilder().evm().erc20(),
+        fee: {
+          // TODO add after merging XCM Payment API rework
+          amount: 0.005,
+          asset: eurc,
+        },
+      },
+      extrinsic: ExtrinsicBuilder().xTokens().transfer(),
     },
   ],
 });